Separate csv out into file_csv.go, make parallel encoding always true on database
