+++ runner/file.go
@@ -2,7 +2,6 @@ package runner
 
 import (
 	"bufio"
-	"encoding/csv"
 	"io"
 	"os"
 	"os/user"
@@ -11,7 +10,6 @@ import (
 	"regexp"
 	"runtime"
 	"strings"
-	"sync"
 
 	"github.com/linkedin/goavro/v2"
 	jsonutil "github.com/multiprocessio/go-json"
@@ -27,108 +25,6 @@ import (
 var preferredParallelism = runtime.NumCPU() * 2
 
 const ParallelEncodingMin = 100 * 1024 * 1024 // 100 MB
-
-func transformCSV(in io.Reader, out io.Writer, delimiter rune, parallelEncoding bool) error {
-	r := csv.NewReader(in)
-	r.Comma = delimiter
-	r.FieldsPerRecord = -1
-	if !parallelEncoding {
-		r.ReuseRecord = true // maybe there is a way to do this when using
-	}
-
-	return withJSONArrayOutWriterFile(out, func(w *jsonutil.StreamEncoder) error {
-		isHeader := true
-		var fields []string
-		row := map[string]any{}
-
-		recordChannel := make(chan []string, preferredParallelism)
-		first := true
-		var wg sync.WaitGroup
-		if parallelEncoding {
-			if _, err := out.Write([]byte{'['}); err != nil {
-				return err
-			}
-			for i := 0; i < preferredParallelism; i++ {
-				go encodeCSVSubroutine(&wg, out, &fields, recordChannel)
-			}
-		}
-
-		for {
-			record, err := r.Read()
-			if err == io.EOF {
-				err = nil
-				break
-			}
-
-			if err != nil {
-				return err
-			}
-
-			if isHeader {
-				for _, field := range record {
-					fields = append(fields, field)
-				}
-				isHeader = false
-				continue
-			}
-
-			if parallelEncoding && !first {
-				recordChannel <- record
-				continue
-			}
-
-			for i, field := range fields {
-				if i < len(record) {
-					row[field] = record[i]
-				} else {
-					row[field] = nil
-				}
-			}
-
-			if parallelEncoding {
-				err = w.SetArray(false).EncodeRow(row) // Encode the first item here so the remaining ones can insert commas freely.
-				if err != nil {
-					return err
-				}
-				w.Close()
-				first = false
-				continue
-			}
-
-			err = w.EncodeRow(row)
-			if err != nil {
-				return err
-			}
-		}
-
-		if parallelEncoding {
-			for {
-				if len(recordChannel) > 0 {
-					continue
-				} else {
-					close(recordChannel)
-					break
-				}
-			}
-			wg.Wait()
-			if _, err := out.Write([]byte{']'}); err != nil {
-				return err
-			}
-		}
-
-		return nil
-	})
-}
-
-func transformCSVFile(in string, out io.Writer, delimiter rune, parallelEncoding bool) error {
-	f, err := os.Open(in)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	return transformCSV(f, out, delimiter, parallelEncoding)
-}
 
 func transformJSON(in io.Reader, out io.Writer) error {
 	_, err := io.Copy(out, in)
@@ -771,30 +667,4 @@ func resolvePath(p string) string {
 	}
 
 	return path.Join(HOME, p[2:])
-}
-
-func encodeCSVSubroutine(wg *sync.WaitGroup, out io.Writer, fields *[]string, recordChannel chan []string) {
-	wg.Add(1)
-	defer wg.Done()
-	w := jsonutil.NewGenericStreamEncoder(out, jsonMarshal, false).SetFirst(false)
-	defer w.Close()
-	row := map[string]any{}
-	var record []string
-	for {
-		record = <-recordChannel
-		if record != nil {
-			for i, f := range *fields {
-				if i >= len(record) {
-					row[f] = nil
-				} else {
-					row[f] = record[i]
-				}
-			}
-			if err := w.EncodeRow(row); err != nil {
-				panic(err) // could we use something like err group and just return the error as normal?
-			}
-		} else {
-			break
-		}
-	}
 }
